+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   KFP SDK version bumped to 1.8.11 in order to fix misbehaving TTL issue
+-   Add Kedro environment name to the pipeline name during upload
 
 ## [0.6.2] - 2022-03-10